# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PagoPa Cruscotto Backend API",
    "description": "PagoPa Cruscotto Backend API documentation",
    "termsOfService": "",
    "license": {
      "name": "unlicensed",
      "url": ""
    },
    "version": "0.6.3"
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/auth-permissions": {
      "get": {
        "tags": [
          "auth-permission-resource"
        ],
        "operationId": "getAllAuthPermissions",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "required": false,
            "schema": {
              "maxLength": 100,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "modulo",
            "in": "query",
            "required": false,
            "schema": {
              "maxLength": 50,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthPermissionDTO"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "auth-permission-resource"
        ],
        "operationId": "updateAuthPermission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthPermissionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthPermissionDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "auth-permission-resource"
        ],
        "operationId": "createAuthPermission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthPermissionDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthPermissionDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth-permissions/{id}": {
      "get": {
        "tags": [
          "auth-permission-resource"
        ],
        "operationId": "getAuthPermission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthPermissionDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "auth-permission-resource"
        ],
        "operationId": "deleteAuthPermission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth-permissions/auth-function/{idFunction}": {
      "get": {
        "tags": [
          "auth-permission-resource"
        ],
        "operationId": "getAllAuthFunctionsSelected",
        "parameters": [
          {
            "name": "idFunction",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthPermissionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth-permissions/auth-function/{idFunction}/associabili": {
      "get": {
        "tags": [
          "auth-permission-resource"
        ],
        "operationId": "getAllAuthFunctionsAssociabili",
        "parameters": [
          {
            "name": "idFunction",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "nameFilter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthPermissionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth-functions/auth-group/{idGroup}/associabili": {
      "get": {
        "tags": [
          "auth-function-resource"
        ],
        "operationId": "getAllAuthFunctionsAssociabili_1",
        "parameters": [
          {
            "name": "idGroup",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "nameFilter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthFunctionDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthPermissionDTO": {
        "required": [
          "modulo",
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "modulo": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "AuthFunctionDTO": {
        "required": [
          "descrizione",
          "modulo",
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "modulo": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "descrizione": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "groupId": {
            "type": "integer",
            "format": "int64"
          },
          "selected": {
            "type": "boolean"
          },
          "authPermissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthPermissionDTO"
            }
          }
        }
      },
...
    }
  }
}